backend/vision.py
from config import llm_vision
from langchain_core.messages import HumanMessage
import base64

# NOTE: In production, you would use a library to convert PDF pages to images first.
# This function assumes the input is processed or mocked for the MVP.
def analyze_charts(pdf_path):
    # Mocking the vision response for stability unless you have 'pdf2image' set up.
    # To make this real, you need 'poppler' installed on your system.
    return """
    {
        "chart_data": {
            "accuracy": "98.5%",
            "latency": "12ms",
            "comparison": "Model A is 15% faster than Model B"
        }
    }
    """